refactor: adjust components
