+++ src/components/goals/CheckGoal/CheckGoalModal.tsx
@@ -1,5 +1,5 @@
 import { ICheckGoal } from "@/interfaces/goals/checkGoals/iCheckGoal";
-import { InputField } from "@/components/InputField";
+import { InputField } from "@/components/Inputs/InputField";
 import Modal from "@/components/Modal";
 import { SetStateAction, useState } from "react";
 import Swal from "sweetalert2";
