+++ src/components/goals/ProgressGoal/ProgressGoalModal.tsx
@@ -1,5 +1,5 @@
 import { IProgressGoal } from "@/interfaces/goals/progressGoals/iProgressGoal";
-import { InputField } from "@/components/InputField";
+import { InputField } from "@/components/Inputs/InputField";
 import Modal from "@/components/Modal";
 import { SetStateAction, useState } from "react";
 import Swal from "sweetalert2";
